The TimeStamp class is made as public.
